[audio] refactored AnoopamAudio service request call
